myapp/requirements.txt
mysql-connector-python
Flask
#Flask-MySQLdb